C code cleanup, per-frame mouse state uploading

+++ runtime/scripts/ui_controls.lua
@@ -3,7 +3,7 @@ ffi.cdef
 [[
     int ui_drawNode(float x, float y, float w, float h, int widget_state, const char* title, char r, char g, char b, char a);
     void ui_dbgTextPrintf(int y, const char *str);
-    int ui_getKeyboardState(uint16_t key);
+    uint8_t ui_getKeyboardState(uint16_t key);
 ]]
 
 local drawNode = ffi.C.ui_drawNode
